# Optimize bundler performance