This refactors the data structures in the default bundler and bundle graph to optimize for performance. This is the result of a lot of profiling, but the main optimizations are:

* Store the `reachableRoots` data structure as an array of bit sets instead of a graph. For each asset, there is a bit set that stores which bundle roots are reachable, along with an inverse mapping. This enables much faster set operations without the cost of hashing, along with bulk operations like intersection. This uses `2 * numBundleRoots * numAssets / 8` bytes of space. For an extremely huge project that might be significant, but should still be in the 10s of megabytes.
* A faster bit set implementation. The previous one used BigInt, which is not terribly well optimized by v8. The new one uses a fixed size `Uint32Array` defined at construction time, and a tiny wasm program exposing the [count trailing zeros](https://developer.mozilla.org/en-US/docs/WebAssembly/Reference/Numeric/Count_trailing_zeros) instruction to JS for iteration over the set bits.
* Storing graph nodes as an array rather than a hash map, since the keys are numbers. This allows for faster lookups without hashing. Since node ids are never reused when a node is deleted, this also works, however, it will leave behind null values in the array. Not sure if this will be a problem.
* A fast path depth-first search implementation for the common case of an enter visitor over all edge types. This avoids recursion, and is instead implemented using a stack. It also uses a faster method of iterating over child edges of a node that doesn't allocate memory to store the ids in an array, and a bit set to track the visited nodes. This avoids the abstraction cost of passing in a custom function to get the next node and filter by edge type. If these pre-conditions are not met, it will fall back to the previous slower implementation.

## Benchmarks

From the React Spectrum docs:

**Before:**
Cold build: 60.31s
Rebuild (adding dependency): 22.32s

**After:**
Cold build: 39.42s
Rebuild (adding dependency): 2.98s

Those are full builds, including transforming and packaging. The bundling phase alone went from 21.19s to 1.49s, or almost 15x faster. 🚀